# docs/endpoints/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ithina Admin Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tenants/{tenant_id}/org-tree": {
      "get": {
        "tags": [
          "org-tree"
        ],
        "summary": "Get organisation tree for a tenant",
        "description": "Returns the tenant's org tree. Smart-default behavior: small tenants (<=500 ACTIVE non-TENANT nodes) get the full tree; larger tenants get a depth-limited tree (default depth=4) with deeper nodes available via the children endpoint. Each returned node carries `has_children`, `child_count`, and `loaded_children` so the frontend knows which subtrees to lazy-fetch. If the depth-limited tree still exceeds 1000 nodes, the server auto-reduces depth and sets `truncated=true`.",
        "operationId": "get_org_tree_api_v1_tenants__tenant_id__org_tree_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 6,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional. Max depth of nodes returned (depth from the implicit TENANT root). 1 = HQ only; 4 = HQ + Country + Region + Store. If omitted, the server picks based on tenant size.",
              "title": "Depth"
            },
            "description": "Optional. Max depth of nodes returned (depth from the implicit TENANT root). 1 = HQ only; 4 = HQ + Country + Region + Store. If omitted, the server picks based on tenant size."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgTreeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "org-tree"
        ],
        "summary": "Add an org_node under an existing parent",
        "description": "Multi-audience write (Step 6.13). Gated on `ADMIN.ORG_NODES.CONFIGURE.TENANT`. SUPER_ADMIN and PLATFORM_ADMIN pass via the GLOBAL->TENANT cascade; OWNER passes via direct TENANT grant. Cross-tenant calls by TENANT users surface as 404 (RLS-as-404 per D-17). The new node's ltree path is parent.path || lower(code).replace('-', '_'). Cascade-order rule: parent's node_type must sit strictly above the child's in the canonical sequence TENANT -> BUSINESS_UNIT -> HQ -> COUNTRY -> REGION -> STORE -> DEPARTMENT; level skipping is allowed.",
        "operationId": "add_org_node_api_v1_tenants__tenant_id__org_tree_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgNodeCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgNodeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/org-nodes/{node_id}/children": {
      "get": {
        "tags": [
          "org-tree"
        ],
        "summary": "Get immediate children of an org-node (lazy-load)",
        "description": "Returns the immediate ACTIVE children of `node_id` within `tenant_id`. Used by the frontend to lazy-load subtrees that were not included in the initial /org-tree response (nodes with `has_children=true` and `loaded_children='none'`). Paginated by offset/limit. Each child carries its own `has_children` and `child_count` for further lazy expansion. Each child's `loaded_children` is 'none' (lazy by default — caller invokes this endpoint again for grandchildren if needed).",
        "operationId": "get_node_children_api_v1_tenants__tenant_id__org_nodes__node_id__children_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Node Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Pagination limit (max 200; default 100).",
              "default": 100,
              "title": "Limit"
            },
            "description": "Pagination limit (max 200; default 100)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgNodeChildrenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/org-tree/{node_id}": {
      "patch": {
        "tags": [
          "org-tree"
        ],
        "summary": "Rename, recode, or reparent an existing org_node",
        "description": "Multi-audience write (Step 6.13). Same gate as POST. Body carries any of `name`, `code`, `parent_id`; at least one is required (empty body returns 422). `node_type` is immutable. On reparent the node and every descendant under it have their ltree paths rewritten in one transaction. Role assignments anchored at the moved node remain intact (D-11: reference is by stable id, not path). The tenant-root org_node can be renamed or recoded but not reparented.",
        "operationId": "edit_org_node_api_v1_tenants__tenant_id__org_tree__node_id__patch",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Node Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgNodePatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgNodeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrgTreeResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id",
            "description": "Echo of path-param tenant_id."
          },
          "tenant_name": {
            "type": "string",
            "title": "Tenant Name",
            "description": "Current tenants.name. Saves the frontend a cross-lookup when rendering the right-pane header."
          },
          "tenant_root_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Root Id",
            "description": "UUID of the tenant-root org_node (the implicit TENANT-typed node that anchors the tenant's tree). Use this as ``parent_id`` when calling POST /org-tree to create a top-level node directly under the tenant. Distinct from ``tenant_id`` (which is the ``tenants.id`` UUID); the two are independent."
          },
          "tenant_root_code": {
            "type": "string",
            "title": "Tenant Root Code",
            "description": "Code of the tenant-root org_node (matches ``org_nodes.code``). Typically derived from the tenant's display_code or name at tenant-create time."
          },
          "tenant_root_path": {
            "type": "string",
            "title": "Tenant Root Path",
            "description": "ltree path of the tenant-root org_node (single label; matches ``org_nodes.path::text``). All descendants' paths have this as their root segment."
          },
          "stats": {
            "$ref": "#/components/schemas/OrgTreeStats"
          },
          "tree": {
            "items": {
              "$ref": "#/components/schemas/OrgNodeTreeItem"
            },
            "type": "array",
            "title": "Tree",
            "description": "Top-level org nodes (children of the tenant root, whose id/code/path now appear as separate top-level fields on this response). Empty list for tenants with zero descendants beyond the tenant root."
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "tenant_name",
          "tenant_root_id",
          "tenant_root_code",
          "tenant_root_path",
          "stats",
          "tree"
        ],
        "title": "OrgTreeResponse",
        "description": "Response for E2: GET /api/v1/tenants/{tenant_id}/org-tree.\n\nDeliberate D-30 exception: this is a singleton resource for the\ntenant, not a paginatable collection, so the envelope is\n``{tenant_id, tenant_name, tenant_root_id, tenant_root_code,\ntenant_root_path, stats, tree}`` rather than ``{items, pagination}``."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}